Move 3D geometry to Video
