+++ src/Video/Geometry/Geometry3D.cpp
@@ -1,5 +1,6 @@
 #include "Geometry3D.hpp"
 
+using namespace Video;
 using namespace Geometry;
 
 Geometry3D::~Geometry3D() {
